backend/requirements.txt
fastapi
uvicorn
requests
pydantic
uuid
supabase
python-dotenv 